Reverse proxy approach with loadbalancer
Implementation to handle multiple request from different users and send it to different instances in EC2. IP adresses of each instance should be set in main function. In a future commit this is going to be available in a config file. There is an issue when sending request from Postman for Desktop that might be fixed.